+++ experiments/train/solo8v2vanilla-ppo2.py
@@ -5,7 +5,7 @@
 #       extension: .py
 #       format_name: percent
 #       format_version: '1.3'
-#       jupytext_version: 1.9.1
+#       jupytext_version: 1.10.0
 #   kernelspec:
 #     display_name: Python 3
 #     language: python
@@ -30,13 +30,10 @@ else:
 # ## Define Experiment Tags
 
 # %%
-TAGS = ['solov2vanilla', 'gpu', 'home_pos_task']
+TAGS = ['solov2vanilla', 'gpu', 'standing_task']
 
 # %% [markdown]
-# ## Get Solo Environment Configuration
-
-# %% [markdown]
-# Import the relevant libraries + rewards & observations
+# # Import required libraries
 
 # %%
 from gym_solo.envs import solo8v2vanilla
@@ -46,12 +43,6 @@ from gym_solo.core import termination as terms
 
 import gym
 import gym_solo
-
-# %% [markdown]
-# Create the config for the enviornment
-
-# %%
-env_config = solo8v2vanilla.Solo8VanillaConfig()
 
 # %% [markdown]
 # ## Parse CLI arguments and register w/ wandb
@@ -64,19 +55,37 @@ from auto_trainer import params
 import auto_trainer
 
 # %% [markdown]
-# Create a basic config. Give the robot a total of 60 seconds simulation time to learn how to stand.
+# Give the robot a total of 10 seconds simulation time to learn how to stand.
 
 # %%
-config = params.BaseParameters().parse()
+episode_length = 10 / solo8v2vanilla.Solo8VanillaConfig.dt
+episode_length
 
-config.episodes = 750000
-config.episode_length = 10 / env_config.dt
+# %% [markdown]
+# Create a basic config
 
+# %%
+config = params.WandbParameters().parse()
+
+config.episodes = 800
+config.episode_length = episode_length
+
+config.num_workers = 6
+config.eval_frequency = 10
+config.eval_episodes = 3
+config.fps = 15
+
+# Create a 3 second gif
+config.eval_render_freq = int(config.episode_length / (3 * config.fps))
+
+config
+
+# %%
 config, run = auto_trainer.get_synced_config(config, TAGS)
 config
 
+
 # %% [markdown]
-# ## Setup Environment
 # Add the following inputs to the robot / environment:
 #
 # **Observations**
@@ -88,23 +97,51 @@ config
 #
 # **Termination Criteria**
 # - Terminate after $n$ timesteps
+#
+# Note that the autotrainer requires that the training environment be a `VecEnv` and the testing environment be a standard `gym.Env` for multi-processing.
+#
+# For us personally, we find that the easiest way to handle this is to create a Stable Baselines `VecEnv` generator (example can be found [here](https://stable-baselines.readthedocs.io/en/master/guide/examples.html#multiprocessing-unleashing-the-power-of-vectorized-environments)) and use that to generate both the training and testing environments.
+#
+# We also like to link our generator with our W&B config so that we can dynamically change the environments based from the web interface. 
+#
+# A full example can be seen below:
 
 # %%
-env = gym.make('solo8vanilla-v0', config=env_config)
+def make_env(length):
+    def _init():
+        env_config = solo8v2vanilla.Solo8VanillaConfig()
+        env = gym.make('solo8vanilla-v0', config=env_config)
 
-env.obs_factory.register_observation(obs.TorsoIMU(env.robot))
-env.obs_factory.register_observation(obs.MotorEncoder(env.robot))
+        env.obs_factory.register_observation(obs.TorsoIMU(env.robot))
+        env.obs_factory.register_observation(obs.MotorEncoder(env.robot))
 
-# env.reward_factory.register_reward(1, rewards.UprightReward(env.robot))
-env.reward_factory.register_reward(1, rewards.HomePositionReward(env.robot))
+        env.reward_factory.register_reward(1, rewards.UprightReward(env.robot))
+        env.termination_factory.register_termination(terms.TimeBasedTermination(length))
+        return env
+    return _init
 
-env.termination_factory.register_termination(terms.TimeBasedTermination(config.episode_length))
+
+# %% [markdown]
+# ### Create the Envs
+# Import the desired vectorized env
+
+# %%
+from stable_baselines.common.vec_env import SubprocVecEnv
+
+# %% [markdown]
+# Create training & testing environments
+
+# %%
+train_env = SubprocVecEnv([make_env(config.episode_length) 
+                           for _ in range(config.num_workers)])
+test_env = make_env(config.episode_length)()
 
 # %% [markdown]
 # ## Learning
+# And we're off!
 
 # %%
-model, config, run = auto_trainer.train(env, config, TAGS, log_freq=100,
-                                        full_logging=False, run=run)
+model, config, run = auto_trainer.train(train_env, test_env, config, TAGS, 
+                                        log_freq=100, full_logging=False, run=run)
 
 # %%
